[#327] refactor : IOS 출처 링크 관련수정
